requirements.txt
alembic==1.6.5
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpickle==2.0.0
Mako==1.1.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1

pandas~=1.3.0
mysql~=0.0.3